pretrain/requirements.txt
absl-py
tensorflow==2.3.1
transformers==4.11.3
ko_lm_dataformat>=0.1.0
